% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merModgANOVA_method.R
\docType{methods}
\name{anova,merModgANOVA-method}
\alias{anova,merModgANOVA-method}
\title{anova method for merModgANVOA}
\usage{
\S4method{anova}{merModgANOVA}(object, ..., ddf = "Satterthwaite", type = 3)
}
\arguments{
\item{object}{an object of class merModgANOVA.}

\item{...}{futher argument. See details.}

\item{ddf}{a character string indicating the method. Default is \code{"Satterthwaite"} and \code{"Kenward-Roger"} is available. See details.}

\item{type}{a integer indicating the equivalent type of SS. Default is 3.}
}
\description{
modified anova method from lmerTest package
}
\details{
See the lmerTest package for more informations.
}
