% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.lmerModgANOVA.R
\name{anova.lmerModgANOVA}
\alias{anova.lmerModgANOVA}
\title{anova method for merModgANVOA}
\usage{
\method{anova}{lmerModgANOVA}(
  object,
  ...,
  type = c("III", "II", "I", "3", "2", "1"),
  ddf = c("Satterthwaite", "Kenward-Roger", "lme4")
)
}
\arguments{
\item{object}{an object of class lmerModgANOVA. See \code{lmerTest} package.}

\item{...}{futher argument. See \code{lmerTest} package.}

\item{type}{a integer indicating the equivalent type of SS. See \code{lmerTest} package.}

\item{ddf}{a character string indicating the method.See \code{lmerTest} package.}
}
\description{
Mdified anova method from the \code{lmerTest} package
}
\seealso{
Other output: 
\code{\link{ranef.lmerModgANOVA}()},
\code{\link{summary.lmerModgANOVA}()}
}
\concept{output}
