% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gANOVA.R
\name{gANOVA}
\alias{gANOVA}
\title{Linear Mixed Models with same variance for contrasts.}
\usage{
gANOVA(formula, data = NULL, REML = TRUE, control = lmerControl(),
  start = NULL, verbose = 0L, subset, weights, na.action, offset,
  contrasts = NULL, devFunOnly = FALSE, ...)
}
\arguments{
\item{formula}{a \code{lme4} formula. If you provide a formula of type \code{(1|id:g)} where \code{id} is the grouping variable and \code{g} is a factor then the covariance structure will estimate the same variance for all orthonormal (\code{contr.poly}) contrasts in \code{g}. WARNINGS The identifier of the grouping variable \code{id} must be written as the first terms to the right of the \code{"|"} because all other terms will be "reduced" by an orthonormal contrasts. See Details for the reduced notation.}

\item{data}{a data frame. See \code{\link{lmer}} for more details.}

\item{REML}{a logical that indicate which criterion to optimize. See \code{\link{lmer}} for more details.}

\item{control}{Some parameters. See \link{lmerControl} or \code{\link{lmer}} for more details.}

\item{start}{starting values for the paramters. See \code{\link{lmer}} for more details.}

\item{verbose}{See \code{\link{lmer}} for more details.}

\item{subset}{an expression to selecte a subset of the data. See \code{\link{lmer}} for more details.}

\item{weights}{an optional vector of weights. See \code{\link{lmer}} for more details.}

\item{na.action}{a function that handle \code{NA}'s. See \code{\link{lmer}} for more details.}

\item{offset}{specify a priori component in the predictor. See \code{\link{lmer}} for more details.}

\item{contrasts}{a list of contrasts. See \code{\link{lmer}} for more details.}

\item{devFunOnly}{a logical set by default to \code{FALSE}. See \code{\link{lmer}} for more details.}

\item{...}{addition arguments. See \code{\link{lmer}} for more details.}
}
\description{
This function provide a small modification of the \code{lme4} function \code{link{lmer}} in order to estimated mixed models with the same variance for orthonormal contrasts. Check formula parameter for details.
}
\details{
\code{summary} and \code{anova} method are copied from the \code{lmerTest} package in order to have p-values.
Given 2 factors, \code{f} and \code{g}, and a grouping variable \code{id}, the formula \code{(1|id) + (1|id:f)+ (1|id:g)+ (1|id:f:g)} can be reduced to  \code{(1|id|f*g)}. However the grouping variable (herer \code{id}) should be written as one variable.
}
\seealso{
\code{\link{lmer}}.
}
