% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.lmerModgANOVA.R
\name{ranef.lmerModgANOVA}
\alias{ranef.lmerModgANOVA}
\title{ranef method for lmerModgANOVA object}
\usage{
ranef.lmerModgANOVA(
  object,
  condVar = FALSE,
  drop = FALSE,
  whichel = names(ans),
  postVar = FALSE,
  ...
)
}
\arguments{
\item{object}{a lmerModgANOVA object.}

\item{condVar}{see lmer.}

\item{drop}{see lmer.}

\item{whichel}{see lmer.}

\item{postVar}{see lmer.}

\item{...}{see lmer.}
}
\description{
Extract the modes of random effects (BLUP) of the model. Just small changes from the ranef.merMod method of lme4 to handle lmerModgANOVA.
}
\seealso{
Other output: 
\code{\link{anova.lmerModgANOVA}()},
\code{\link{summary.lmerModgANOVA}()}
}
\concept{output}
