% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.merModgANOVA.R
\name{ranef.merModgANOVA}
\alias{ranef.merModgANOVA}
\title{ranef method for merModgANOVA object}
\usage{
ranef.merModgANOVA(object, condVar = FALSE, drop = FALSE,
  whichel = names(ans), postVar = FALSE, ...)
}
\arguments{
\item{object}{a merModgANOVA object.}

\item{condVar}{see lmer.}

\item{drop}{see lmer.}

\item{whichel}{see lmer.}

\item{postVar}{see lmer.}

\item{...}{see lmer.}
}
\description{
Extract the modes of random effects (BLUP) of the model. Just small changes from the ranef.merMod method of lme4 to handle merModgANOVA.
}
