% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.lmerModgANOVA.R
\name{refit.lmerModgANOVA}
\alias{refit.lmerModgANOVA}
\title{Refit a lmerModgANOVA model with a new response}
\usage{
refit.lmerModgANOVA(
  object,
  newresp = NULL,
  rename.response = FALSE,
  maxit = 100L,
  ...
)
}
\arguments{
\item{object}{a lmerModgANOVA model}

\item{newresp}{a vector of new responses.}

\item{rename.response}{Replacing the name of in the forumla. Default is FALSE}

\item{maxit}{a integer, not use. Default is 100.}

\item{...}{orther arguments}
}
\description{
Refit a lmerModgANOVA model with a new response
}
\seealso{
\code{\link{refit}}
}
