% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merModgANOVA_method.R
\docType{methods}
\name{summary,merModgANOVA-method}
\alias{summary,merModgANOVA-method}
\title{summary method for merModgANVOA}
\usage{
\S4method{summary}{merModgANOVA}(object, ddf = "Satterthwaite", ...)
}
\arguments{
\item{object}{an object of class merModgANOVA.}

\item{ddf}{a character string indicating the method. Default is \code{"Satterthwaite"} and \code{"Kenward-Roger"} is available. See details.}

\item{...}{futher argument. See details.}
}
\description{
modified summary method from lmerTest package
}
\details{
See the lmerTest package for more informations.
}
