% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lmerModgANOVA.R
\name{summary.lmerModgANOVA}
\alias{summary.lmerModgANOVA}
\title{summary method for merModgANVOA}
\usage{
\method{summary}{lmerModgANOVA}(object, ..., ddf = c("Satterthwaite", "Kenward-Roger", "lme4"))
}
\arguments{
\item{object}{an object of class lmerModgANOVA.}

\item{...}{futher argument. See details.}

\item{ddf}{a character string indicating the method. Default is \code{"Satterthwaite"} and \code{"Kenward-Roger"} is available. See details.}
}
\description{
modified summary method from lmerTest package
}
\details{
See the lmerTest package for more informations.
}
\seealso{
Other output: 
\code{\link{anova.lmerModgANOVA}()},
\code{\link{ranef.lmerModgANOVA}()}
}
\concept{output}
